database migration with checks

+++ database/migrations/2016_06_16_144537_add_branding_column_to_form_mail_table.php
@@ -12,9 +12,11 @@ class AddBrandingColumnToFormMailTable extends Migration
      */
     public function up()
     {
-        Schema::table('form_mail', function (Blueprint $table) {
-            $table->addColumn('text', 'branding');
-        });
+        if (Schema::hasTable('form_mail') and !Schema::hasColumn('form_mail','branding' )) {
+            Schema::table('form_mail', function (Blueprint $table) {
+                $table->addColumn('text', 'branding');
+            });
+        }
     }
 
     /**
@@ -24,8 +26,10 @@ class AddBrandingColumnToFormMailTable extends Migration
      */
     public function down()
     {
-        Schema::table('form_mail', function (Blueprint $table) {
-            $table->dropColumn('branding');
-        });
+        if (Schema::hasTable('form_mail') and Schema::hasColumn('form_mail','branding' )) {
+            Schema::table('form_mail', function (Blueprint $table) {
+                $table->dropColumn('branding');
+            });
+        }
     }
 }
